adds two audio files, a long space between words, and a short space between letters. Merges sandbox.py with main.py, implemented feature of playing morse code when clicking play message button after translation.
