+++ main.py
@@ -1,4 +1,6 @@
 from tkinter import *
+import pygame
+import time
 
 morse_text_translation = {
   "a": ".-",
@@ -49,11 +51,11 @@ def entry_validation(event=None):
 
     user_input = entry.get().strip().lower()
     if all(char.isalpha() or char.isspace() for char in user_input):
-        translate(user_input.split())  # Split the input by spaces and passes the list to translate function
+        translate(user_input.split(), user_input)  # Split the input by spaces and passes the list to translate function
     else:
         instruction_label.config(text="Please enter only letters and spaces.")
 
-def translate(user_input_list):
+def translate(user_input_list, user_input):
     global audio_button_created
     for word in user_input_list:
         for x in range(len(word)):
@@ -62,12 +64,28 @@ def translate(user_input_list):
             morse_list.append('/')
     morse_code_result.config(text=f'{"  ".join(morse_list)}')
     if not audio_button_created:
-        audio_button = Button(text="Play message", font=("Arial", 15), command=play_audio)
+        audio_button = Button(text="Play message", font=("Arial", 15), command=lambda:play_user_message(user_input))
         audio_button.pack()
         audio_button_created = True
 
-def play_audio():
-    pass
+def play_user_message(user_input):
+    for letter in user_input:
+        if letter == " ":
+            letter = "space"
+        audio_file = f"text-to-morse-code/audio/{letter}.mp3"  # Replace with the path to your audio file
+        play_audio(audio_file)
+        small_space = "text-to-morse-code/audio/letter_space.mp3" # adds a little period of silence to differentiate between start and end of letters
+        play_audio(small_space)
+
+def play_audio(audio_file):
+    pygame.mixer.music.load(audio_file)  # Load the audio file
+    pygame.mixer.music.play()  # Play the audio
+    while pygame.mixer.music.get_busy():
+        # Keep the program running while the audio is playing
+        time.sleep(0.1)
+
+# Initialize Pygame (should be done once at the beginning of your script)
+pygame.mixer.init()
 
 entry.bind("<Return>", entry_validation)
